Moved all of the code of the purchase section to the Wrapper class
