+++ src/components/screens/PurchaseSection.js
@@ -20,9 +20,9 @@ class PurchaseSection extends React.Component {
             totalPrice: 0.0
         };
         // this.addItems = this.addItems.bind(this);
-        this.changeQuantityEvent = this.changeQuantityEvent.bind(this);
-        this.getCartItems = this.getCartItems.bind(this);
-        this.removeCartItems = this.removeCartItems.bind(this);
+        // this.changeQuantityEvent = this.changeQuantityEvent.bind(this);
+        // this.getCartItems = this.getCartItems.bind(this);
+        // this.removeCartItems = this.removeCartItems.bind(this);
         this.onSaveForLater = this.onSaveForLater.bind(this);
     }
 
@@ -67,80 +67,80 @@ class PurchaseSection extends React.Component {
         //     });
     }
 
-    changeQuantityEvent = () =>  {
-        console.log("I RAN");
-        this.props.changeQuantity(item, event)
-    }
+    // changeQuantityEvent = () =>  {
+    //     console.log("I RAN");
+    //     this.props.changeQuantity(item, event)
+    // }
 
-    getCartItems(books) {
-        let total = 0.0;
-        console.log("BOOKS START:")
-        console.log(books);
-        //Below all books get mapped onto the cart. Delete after
-        let cart = books.map(item => {
-            total += item.price * item.quantity;
-            return (
-                <tr key={item.bookid}>
-                    <td>{item.title}</td>
-                    <td>
-                        x
-                        <input
-                            class='purchase-input'
-                            type='number'
-                            value={item.quantity}
-                            onChange={this.changeQuantityEvent.bind(this.item,this)}
-                        />
-                    </td>
-                    <td>${item.price.toFixed(2)}</td>
+    // getCartItems(books) {
+    //     let total = 0.0;
+    //     console.log("BOOKS START:")
+    //     console.log(books);
+    //     //Below all books get mapped onto the cart. Delete after
+    //     let cart = books.map(item => {
+    //         total += item.price * item.quantity;
+    //         return (
+    //             <tr key={item.bookid}>
+    //                 <td>{item.title}</td>
+    //                 <td>
+    //                     x
+    //                     <input
+    //                         class='purchase-input'
+    //                         type='number'
+    //                         value={item.quantity}
+    //                         onChange={this.changeQuantityEvent.bind(this.item,this)}
+    //                     />
+    //                 </td>
+    //                 <td>${item.price.toFixed(2)}</td>
 
-                    <td>
-                        <button
-                            type='button'
-                            class='btn btn-outline-dark'
-                            onClick={this.onSaveForLater.bind(this, item)}
-                        >
-                            Save For Later
-                        </button>
-                        <Button
-                            onClick={this.removeCartItems.bind(this, item)}
-                            style={{
-                                backgroundColor: "rgba(0,0,0,0)",
-                                border: "none"
-                            }}
-                        >
-                            <Icon name='close' color='red' />
-                        </Button>
-                    </td>
-                </tr>
-            );
-        });
-        console.log(books);
-        this.setState({ cartBooks: books, cartItems: cart, totalPrice: total });
-        this.forceUpdate();
-    }
+    //                 <td>
+    //                     <button
+    //                         type='button'
+    //                         class='btn btn-outline-dark'
+    //                         onClick={this.onSaveForLater.bind(this, item)}
+    //                     >
+    //                         Save For Later
+    //                     </button>
+    //                     <Button
+    //                         onClick={this.removeCartItems.bind(this, item)}
+    //                         style={{
+    //                             backgroundColor: "rgba(0,0,0,0)",
+    //                             border: "none"
+    //                         }}
+    //                     >
+    //                         <Icon name='close' color='red' />
+    //                     </Button>
+    //                 </td>
+    //             </tr>
+    //         );
+    //     });
+    //     console.log(books);
+    //     this.setState({ cartBooks: books, cartItems: cart, totalPrice: total });
+    //     this.forceUpdate();
+    // }
 
-    removeCartItems(item) {
-        this.state.cartBooks.splice(
-            this.state.cartBooks.findIndex(book => book.bookID === item.bookid),
-            1
-        );
-        this.getCartItems(this.state.cartBooks);
-        fetch("/cart/delete", {
-            method: "POST",
-            body: JSON.stringify({
-                userid: this.props.username,
-                bookid: item.bookid
-            }),
-            headers: { "Content-Type": "application/json" }
-        })
-            .then(res => console.log(res.body))
-            .then(newInfo => {
-                console.log("Item Deleted");
-            })
-            .catch(err => {
-                console.log(err);
-            });
-    }
+    // removeCartItems(item) {
+    //     this.state.cartBooks.splice(
+    //         this.state.cartBooks.findIndex(book => book.bookID === item.bookid),
+    //         1
+    //     );
+    //     this.getCartItems(this.state.cartBooks);
+    //     fetch("/cart/delete", {
+    //         method: "POST",
+    //         body: JSON.stringify({
+    //             userid: this.props.username,
+    //             bookid: item.bookid
+    //         }),
+    //         headers: { "Content-Type": "application/json" }
+    //     })
+    //         .then(res => console.log(res.body))
+    //         .then(newInfo => {
+    //             console.log("Item Deleted");
+    //         })
+    //         .catch(err => {
+    //             console.log(err);
+    //         });
+    // }
 
     saveForLater() {}
